requirements.txt
asttokens==3.0.0
cachetools==6.2.1
click==8.3.0
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.3
-e git+ssh://git@github.com/trungdung1711/credit_card_fraud_detection.git@bb12b6dee8b127c76f1a50cb1f0e17e467bcbe6a#egg=credit_card_fraud_detection
cuda-bindings==13.0.3
cuda-core==0.3.2
cuda-pathfinder==1.3.1
cuda-python==13.0.3
cuda-toolkit==13.0.1
cudf-cu13==25.10.0
cuml-cu13==25.10.0
cupy-cuda13x==13.6.0
cycler==0.12.1
dask==2025.9.1
dask-cuda==25.10.0
dask-cudf-cu13==25.10.0
debugpy==1.8.17
decorator==5.2.1
distributed==2025.9.1
distributed-ucxx-cu13==0.46.0
executing==2.2.1
fastrlock==0.8.3
fonttools==4.60.1
fsspec==2025.9.0
imbalanced-learn==0.14.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
libcudf-cu13==25.10.0
libcuml-cu13==25.10.0
libkvikio-cu13==25.10.0
libraft-cu13==25.10.0
librmm-cu13==25.10.0
libucx-cu13==1.19.0
libucxx-cu13==0.46.0
llvmlite==0.44.0
locket==1.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.2
nest-asyncio==1.6.0
numba==0.61.2
numba-cuda==0.19.1
numpy==2.2.6
nvidia-cublas==13.0.2.14
nvidia-cuda-cccl==13.0.85
nvidia-cuda-crt==13.0.88
nvidia-cuda-nvcc==13.0.88
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cufft==12.0.0.61
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.28.3
nvidia-nccl-cu13==2.28.7
nvidia-nvjitlink==13.0.88
nvidia-nvvm==13.0.88
nvtx==0.2.13
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.2.250926
parso==0.8.5
partd==1.4.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
Pygments==2.19.2
pylibcudf-cu13==25.10.0
pylibraft-cu13==25.10.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
raft-dask-cu13==25.10.0
rapids-dask-dependency==25.10.0
rapids-logger==0.1.19
rich==14.2.0
rmm-cu13==25.10.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
sortedcontainers==2.4.0
stack-data==0.6.3
tblib==3.2.0
threadpoolctl==3.6.0
toolz==1.1.0
tornado==6.5.2
traitlets==5.14.3
treelite==4.4.1
types-pytz==2025.2.0.20250809
typing_extensions==4.15.0
tzdata==2025.2
ucxx-cu13==0.46.0
urllib3==2.5.0
wcwidth==0.2.14
xgboost==3.0.5
zict==3.0.0
